+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableShortArray.kt
@@ -522,6 +522,17 @@ public value class ImmutableShortArray @PublishedApi internal constructor(
         return temp.toImmutableArray()
     }
 
+    /**
+     * Returns an immutable array containing only the distinct elements from this immutable array.
+     *
+     * Iteration order is preserved and subsequent duplicate elements are ignored.
+     */
+    public fun distinct(): ImmutableShortArray {
+        if (size <= 1) return this
+
+        return values.toSet().toImmutableArray()
+    }
+
     public companion object {
         @PublishedApi
         internal val EMPTY: ImmutableShortArray = ImmutableShortArray(ShortArray(0))